app/main.py
from dotenv import load_dotenv
import os

# Load environment variables first, before any other imports
load_dotenv()

from fastapi import FastAPI, WebSocket
from fastapi.middleware.cors import CORSMiddleware
import uvicorn
from app.api.call_handler import router as call_router

app = FastAPI(title="99phones API", description="Voice Call Processing API with Sarvam AI Integration")

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Include the call_handler router
app.include_router(call_router, prefix="", tags=["calls"])

# Health check endpoint
@app.get("/health")
async def health_check():
    return {"status": "healthy"}

if __name__ == "__main__":
    port = int(os.getenv("PORT", 8000))
    host = os.getenv("HOST", "0.0.0.0")
    uvicorn.run("app.main:app", host=host, port=port, reload=True) 