pyproject.toml
[project]
name = "99phone"
version = "0.1.0"
description = "AI-powered phone assistant with Twilio integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=0.19.0",
    "twilio>=7.0.0",
    "requests>=2.26.0",
    "python-multipart>=0.0.5",
    "websockets>=10.0",
    "pydantic>=1.8.2",
    "aiohttp>=3.8.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "groq>=0.4.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []
